refactor: refactor package references in tests
closes #135
