+++ src/test/java/ci/integration/GithubAPIHandlerTest.java
@@ -1,9 +1,8 @@
-package ci;
+package ci.integration;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import ci.integration.GithubAPIHandler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.Random;
