+++ src/test/java/ci/service/CompilationTest.java
@@ -1,4 +1,4 @@
-package ci;
+package ci.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import ci.service.CompilationService;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
