+++ src/test/java/ci/util/ValidationTest.java
@@ -1,4 +1,4 @@
-package ci;
+package ci.util;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertFalse;
